@@ -1,5 +1,5 @@
 import { Injectable } from "@angular/core";
-import { Observable } from 'rxjs/Observable';
+import { Observable } from 'rxjs';
 import { HoraDia } from "../models/HoraDia";
 import { HttpClient } from "@angular/common/http";
 import { GLOBAL } from "./global";
